Reverts TS conversion and client-side sort change for ErrorGroupOverview
