+++ x-pack/plugins/apm/public/components/app/ErrorGroupOverview/List/__test__/List.test.tsx
@@ -13,6 +13,7 @@ import {
   toJson
   // @ts-ignore
 } from '../../../../../utils/testHelpers';
+// @ts-ignore
 import { ErrorGroupList } from '../index';
 import props from './props.json';
